Home

Particles.jl

A Julia package to get access to the PDG particle data.

Particles.available_catalog_files

Particles.use_catalog_file

Particles.available_catalog_files — Method.

available_catalog_files()

Function to get the available catalog files which are available within the package and returns a list with the absolute filepaths.

Examples

julia> Particles.available_catalog_files()
["/home/foobar/dev/Particles.jl/data/particle2019.csv"]

source

Particles.use_catalog_file — Method.

use_catalog_file(filepath::AbstractString)

This function reads a given catalog file and sets it as reference

Arguments

filepath::AbstractString: filepath to the catalog file

Examples

julia> Particles.use_catalog_file("/home/foobar/dev/Particles.jl/data/particle2019.csv")

source

Acknowledgements

This project uses data files which are part of the particle Python package, developed by the Scikit-HEP group. Many thanks for the permission!