API

Corpuscles.jl

A Julia package to get access to the PDG particle data.

Corpuscles.available_catalog_files

Corpuscles.find_particles_by_name

Corpuscles.use_catalog_file

Corpuscles.available_catalog_files — Method.

available_catalog_files()

Function to get the available catalog files which are available within the package and returns a list with the absolute filepaths.

Examples

julia> Corpuscles.available_catalog_files()
["/home/foobar/dev/Corpuscles.jl/data/particle2019.csv"]

source

Corpuscles.find_particles_by_name — Method.

find_particles_by_name(name::Regex)

Find particles by their name

Arguments

name::Union{Regex, AbstractString}: name search term

Examples

julia> Corpuscles.find_particles_by_name(r"[A-Z]*mma")
1-element Array{Particle,1}:
 Particle(22)

source

Corpuscles.use_catalog_file — Method.

use_catalog_file(filepath::AbstractString)

This function reads a given catalog file and sets it as reference

Arguments

filepath::AbstractString: filepath to the catalog file

Examples

julia> Corpuscles.use_catalog_file("/home/foobar/dev/Corpuscles.jl/data/particle2019.csv")

source