Corpuscles.jl

A Julia package to get access to the PDG particle data.

Corpuscles.PDGID

Corpuscles.available_catalog_files

Corpuscles.particles

Corpuscles.use_catalog_file

Corpuscles.PDGID — Type

PDG IDs consist of 7 digits prefixed by a sign, following the scheme:

+/- N Nr Nl Nq1 Nq2 Nq3 Nj

Those are accessible as fields. There are PDG IDs with more than 7 digits for non-standard particles such as Q-balls. To support those, we follow the implementation in the SciKit-HEP particle Python package, which introduced N8, N9 and N10.

Corpuscles.available_catalog_files — Method

available_catalog_files()

Function to get the available catalog files which are available within the package and returns a list with the absolute filepaths.

Examples

julia> Corpuscles.available_catalog_files()
["/home/foobar/dev/Corpuscles.jl/data/particle2019.csv"]

Corpuscles.particles — Method

particles()

Returns the full list of particles from the currently selected catalog.

Corpuscles.use_catalog_file — Method

use_catalog_file(filepath)

This function reads a given catalog file and sets it as reference

Arguments

filepath::AbstractString: filepath to the catalog file

Examples

julia> Corpuscles.use_catalog_file("/home/foobar/dev/Corpuscles.jl/data/particle2019.csv")